requirements.txt
openai>=1.40.0
langfuse>=2.36.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
numpy>=1.26.0
